Extracted non-BQ code from BQ code.
